Added and fixed Tools and Ingredients lists

+++ src/main/java/group3/brewday/repositories/ToolRepository.java
@@ -3,5 +3,5 @@ package group3.brewday.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import group3.brewday.models.Tool;
 
-public interface ToolRepository extends JpaRepository<Tool, Integer>{
+public interface ToolRepository extends JpaRepository<Tool, Long>{
 }
